优化 CurFrameInfoDetail 内存占用，缩小数据类型
## 修改内容

### 1. CurFrameInfoDetail 结构体优化
- curFramePrepareSeqNum: uint32_t → uint16_t
- curFramePostPrepareSeqNum: uint32_t → uint16_t
- curFrameVsyncId: uint64_t → uint32_t

### 2. RSUniRenderVisitor 类同步修改
- IncreasePrepareSeq(): 返回值 uint32_t → uint16_t
- IncreasePostPrepareSeq(): 返回值 uint32_t → uint16_t
- nodePreparedSeqNum_: uint32_t → uint16_t
- nodePostPreparedSeqNum_: uint32_t → uint16_t

### 3. 类型转换
- 添加 static_cast<uint32_t>() 处理 VSync ID 类型转换

## 优化效果
- 结构体大小减少约 50% (18-24字节 → 10-12字节)
- 提高缓存命中率
- 降低内存占用

## 风险评估
- 序列号范围: 0-65535 (约18分钟@60fps后溢出)
- VSync ID 截断高32位数据

+++ rosen/modules/render_service/core/pipeline/main_thread/rs_uni_render_visitor.cpp
@@ -770,7 +770,7 @@ void RSUniRenderVisitor::UpdateCurFrameInfoDetail(RSRenderNode& node, bool subTr
 {
     if (GetDumpRsTreeDetailEnabled()) {
         auto& curFrameInfoDetail = node.GetCurFrameInfoDetail();
-        curFrameInfoDetail.curFrameVsyncId = HgmCore::Instance().GetVsyncId();
+        curFrameInfoDetail.curFrameVsyncId = static_cast<uint32_t>(HgmCore::Instance().GetVsyncId());
         curFrameInfoDetail.curFrameSubTreeSkipped = subTreeSkipped;
         if (isPostPrepare) {
             curFrameInfoDetail.curFramePostPrepareSeqNum = IncreasePostPrepareSeq();
